+++ beam_time/screens/game/mod.rs
@@ -144,7 +144,7 @@ impl Screen for GameScreen {
                     let (cost, _count) = price(&self.board.tiles, level);
 
                     // Award potential steam achievements
-                    award_campaign_achievements(state, level_meta.id);
+                    award_campaign_achievements(state, level_meta.id, (cost, latency));
 
                     state.mark_level_complete(level.id);
                     // Upload solution to leaderboard server